requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
ddgs>=8.0.0
firecrawl>=4.23.0
aiohttp>=3.8.0
requests>=2.30.0
